% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_fbapp_template}
\alias{create_fbapp_template}
\title{Create templates to import on mobiles devices with Field Book App}
\usage{
create_fbapp_template(
  variable = "",
  type = c("numeric", "categorical", "percent", "date", "boolean", "text", "photo",
    "counter", "multicat", "score", "rust_rating", "audio", "location"),
  defaultValue = "",
  minimum = 0,
  maximum = 100,
  details = "",
  categories = NULL,
  isVisible = "",
  realPosition = "",
  num_eval = 1
)
}
\arguments{
\item{variable}{character. variable name}

\item{type}{character. type of variables. There are 12 types of variables \code{numeric}}

\item{defaultValue}{numeric or character. Value to display on the interface by default.}

\item{minimum}{numeric. lower limit value for numerical traits. Only available for type \code{numeric} variables.}

\item{maximum}{numeric. upper limit value for numerical traits. Only available for type \code{numeric} variables.}

\item{details}{character. Provide details about the variable. For example variable's full name or variable's units}

\item{categories}{A character vector of categories of the variable. Only available for type \code{categorical}}

\item{isVisible}{boolean. If \code{true}, the variables should be visible in the interface, otherwise is \code{false}.}

\item{realPosition}{integer. Position of variable in the table}

\item{num_eval}{integer. Number of evaluation per variable}
}
\description{
Create Field Book App template
}
\author{
Omar Benites
}
