% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rfbapp.R
\name{write_rfbapp}
\alias{write_rfbapp}
\title{Write data to an trt file}
\usage{
write_rfbapp(x, file = NULL)
}
\arguments{
\item{x}{object that has table containing a set of variables}

\item{file}{\code{trt} file name}
}
\description{
\code{write_rfbapp} write a data.frame to an \code{.trt} file (FieldBook App format)
}
